+++ src/categories/helpers/categoryExists.ts
@@ -1,11 +1,11 @@
 import { Category } from "../types/category";
 
 export const categoryNameExists = (categories: Category[], name: string) => {
-  return !!categories.some(
+  return categories.some(
     (category: Category) => category.name.toLowerCase() === name.toLowerCase()
   );
 };
 
 export const categoryUuidExists = (categories: Category[], uuid: string) => {
-  return !!categories.some((category) => category.uuid === uuid);
+  return categories.some((category) => category.uuid === uuid);
 };
